Added functionality to the navbar and shopping cart
Added functionality to the shopping cart. Created Shoppingcart.js, Shoppingcart.module.css. Added context files to shoppingcart.js, added a way to calculate the sum of all cars. Added a way to remove cars from the shopping carts.  Added a function which reduces the total price when a car is removed from the shopping cart